Remove unnecessary type assumption in image decoding
The previous version included a comment stating that the 'decoded' variable is already of type `ImageBuffer<Rgb<u8>, Vec<u8>>`. This has been removed because it was found to be unnecessary and not contributing to the understanding or implementation of the code. The code is now cleaner and easier to read.